requirements.txt
psycopg2-binary;
python-frontmatter;
mistune;
python-dotenv;
Flask;
lxml;
beautifulsoup4;
debugpy;
gunicorn;